requirements.txt
annotated-types==0.7.0
anyio==4.4.0
astroid==3.2.4
blis==0.7.11
build==1.2.1
cachetools==5.4.0
catalogue==2.0.10
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
confection==0.1.5
cowgirl-ai-auto-code==1.5.2
cowgirl-ai-utils==1.2.1
cymem==2.0.8
dill==0.3.8
distlib==0.3.9
distro==1.9.0
docutils==0.21.2
filelock==3.16.1
flake8==7.1.1
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.65.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
identify==2.6.1
idna==3.7
importlib_metadata==8.0.0
iniconfig==2.0.0
isort==5.13.2
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
keyring==25.2.1
langcodes==3.4.0
language_data==1.2.0
logging==0.4.9.6
lxml==5.2.2
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.3.0
murmurhash==1.0.10
mypy-extensions==1.0.0
nh3==0.2.18
nltk==3.8.1
nodeenv==1.9.1
numpy==1.26.4
openai==1.51.2
packaging==24.1
pathspec==0.12.1
pkginfo==1.10.0
platformdirs==4.2.2
pluggy==1.5.0
pre_commit==4.0.1
preshed==3.0.9
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.1
pydantic==2.8.2
pydantic_core==2.20.1
pyflakes==3.2.0
Pygments==2.18.0
pylint==3.2.6
pyproject_hooks==1.1.0
pytest==8.3.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
PyYAML==6.0.2
readme_renderer==44.0
regex==2024.5.15
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
rsa==4.9
setuptools==71.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.2.5
tomlkit==0.13.0
tqdm==4.66.4
twine==5.1.1
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.26.6
wasabi==1.1.3
weasel==0.4.1
wrapt==1.16.0
zipp==3.19.2
